# Conflicts:
#	Reservation/src/main/java/net/travel/reservation/controller/ReservationController.java
#	Reservation/src/main/java/net/travel/reservation/repositories/ReservationRepository.java
#	Reservation/src/main/java/net/travel/reservation/services/ReservationService.java

+++ Reservation/src/main/java/net/travel/reservation/entites/Service.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -24,7 +25,7 @@ public class Service {
     private String nom; // Ex: "DJ", "Décoration", "Traiteur", "Photographe"
     @Column(length = 500)
     private String description;
-    private Double prix; // Prix du service en EUR/TND...
+    private BigDecimal prix; // Prix du service en EUR/TND...
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "salle_id", nullable = false)
     private Salle salle;
